[dry-tx] Introduce an initial DestroyWork transaction [ci-skip]
This `DestroyWork` transaction takes care of the basic process of destroying a
work and its related database objects, similarly to what happens in the Actor
Stack.

The actor stack optimizes somewhat for `FileSet` destruction (see
`CleanupFileSetsActor`) by deleting the `#list_source` and removing the object
from Solr prior to destroying `FileSet`s. We omit these optimizations for the
time being. They can be re-introduced as additional steps, if desired.